Created NodeShapeControlGroup
